switch from starting-guide fragment to guide fragment
