+++ app/src/main/java/kr/mashup/mapc/ui/main/guide/StartingGuideFragment.java
@@ -1,0 +1,67 @@
+package kr.mashup.mapc.ui.main.guide;
+
+import android.os.AsyncTask;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import kr.mashup.mapc.R;
+
+public class StartingGuideFragment extends Fragment {
+
+    View view;
+    Button guide_start_btn;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.fragment_guide, container, false);
+
+        guide_start_btn = view.findViewById(R.id.guide_start_btn);
+        guide_start_btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Fragment fragment = new GuideFragment();
+                FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
+                FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+                fragmentTransaction.replace(R.id.view_pager, fragment);
+                fragmentTransaction.addToBackStack(null);
+                fragmentTransaction.commit();
+            }
+        });
+
+       return view;
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+    }
+
+    public static StartingGuideFragment newInstance() {
+
+        Bundle args = new Bundle();
+
+        StartingGuideFragment fragment = new StartingGuideFragment();
+        fragment.setArguments(args);
+        return fragment;
+    }
+    public class NaverCSSTask extends AsyncTask<String[],Void, String>{
+        @Override
+        protected String doInBackground(String[]... strings) {
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(String s) {
+            super.onPostExecute(s);
+        }
+    }
+}
